+++ src/Phery/PheryException.php
@@ -1,0 +1,43 @@
+<?php
+/**
+ * The MIT License (MIT)
+ *
+ * Copyright © 2010-2013 Paulo Cesar, http://phery-php-ajax.net/
+ *
+ * Permission is hereby granted, free of charge, to any person
+ * obtaining a copy of this software and associated documentation
+ * files (the “Software”), to deal in the Software without restriction,
+ * including without limitation the rights to use, copy, modify, merge,
+ * publish, distribute, sublicense, and/or sell copies of the Software,
+ * and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included
+ * in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS
+ * OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
+ * THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR
+ * OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
+ * ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+ * OTHER DEALINGS IN THE SOFTWARE.
+ *
+ * @link       http://phery-php-ajax.net/
+ * @author     Paulo Cesar
+ * @version    2.7.2
+ * @license    http://opensource.org/licenses/MIT MIT License
+ */
+
+namespace Phery;
+
+use Exception;
+
+/**
+ * Exception class for Phery specific exceptions
+ *
+ * @package    Phery
+ */
+class PheryException extends Exception {
+
+}
